Fix compilation errors for gcc 9.4.0 and older
Observed many compilation errors while compilating by gcc 9.4.0
(Ubuntu 20.04).
Ex:
src/vm.h:521:17: error: a label can only be part of a statement and
a declaration is not a statement

According to previous C standards: forbidden to put a label
immediately before a declaration.
As workaround the empty statement can be added before a label.